Added folder structure and created seperate Cart Card component
